Added features
Expressions now support bitwise AND, OR, and XOR operations. These have lowest precedence of the currently implemented operations ([],*/,+-,&|^). Added string to ASCII byte string conversions to the preprocessor. Strings can now be defined with .BYTE "<str contents>". A null terminator is not automatically inserted, but can be included with a \0 or closing the string and including a ,$00 after. Double quotes and backslashes can also be escaped. For any other special characters, a byte will need to be entered outside of the string.