Added negation and bitwise NOT
New expression operations for the unary operators, negation (-e) and bitwise NOT (~e) have been added. Expressions using these will be treated as half-size (8-bit) expressions if all numbers are under 256 AND all hex numbers are two digits or less. Any 3+ digit hex values will be treated as full-size and will cause the expression to be a full-size expressoin.